feat: Separate service methods into traits
Structure service based cloud methods into dedicated traits to keep code
smaller.
